Changed how time is handled.
The current arduino board i'm using seems to be about 20% fast when
using millis().  For example, 10 seconds that goes by on the GPS only
reads as ~8000 millis.  This could explain why so many calculations for
distance_RPM and speed_RPM were so wonky. Also, this explains why the
old graphs and data would show that the time would "jump ahead" every 5
minutes (since the way we were using the Time.cpp library in conjunction
with the GPS, it would prime the time subsystem with the GPS every 5
minutes).

The time is now being retrieved directly from the GPS and used (as
opposed to using it to set the internal clock and then relying on
millis()).

We're now only using millis() to record the data for future debugging
and also to check for non-critical things like when to flush to disk.

Hopefully this will fix a bunch of things.